Resolved : F_Digits_Summation - I make mistake in int to long long

+++ F_Digits_Summation.cpp
@@ -1,19 +1,19 @@
-#include <bits/stdc++.h>
+#include <iostream>
 using namespace std;
 
 int main()
 {
-    int m, n;
+    long long m, n;
     
     // take input
     cin >> m >> n;
 
     // % 10 to find last digit
-    int m_last = m % 10;
-    int n_last = n % 10;
+    long long m_last = m % 10;
+    long long n_last = n % 10;
 
     // sum
-    int sum = m_last + n_last;
+    long long sum = m_last + n_last;
 
     // print sum
     cout << sum << endl;
